+++ R/flget_bathymetry.R
@@ -31,10 +31,9 @@
 #' @export
 #'
 #' @examples
-#' # Download+load data
+#' # Load data
 #' fjord_code <- "test"
-#' fl_DownloadFjord(fjord_code, dirdata = tempdir())
-#' fjorddata <- fl_LoadFjord(fjord_code, dirdata = tempdir())
+#' fjorddata <- fl_LoadFjord(fjord_code, dirdata = system.file("extdata", package = "FjordLight"))
 #'
 #' # Shallow data (what = "s"; s for shallow) as a data.frame
 #' shallow_df <- flget_bathymetry(fjorddata, what = "s", mode = "df")
@@ -49,7 +48,7 @@
 #' land_only <- flget_bathymetry(fjorddata, what = "l", mode = "raster", PLOT =TRUE)
 #' }
 #'
-#' # For more examples: https://face-it-project.github.io/FjordLight/articles/fl_example.html
+#' # For more examples see: https://face-it-project.github.io/FjordLight/articles/fl_example.html
 #'
 flget_bathymetry <- function(fjord,
                              what = "o",