+++ tests/testthat/test-fl_ListFjords.R
@@ -1,8 +1,7 @@
 # Testing for fl_ListFjords
 
 test_that("fl_ListFjords works", {
+  skip_if_offline()
   list_test <- fl_ListFjords()
   expect_type(list_test, "character")
-  expect_equal(length(list_test), 8)
-  expect_equal(list_test[2], "is")
 })