+++ tests/testthat/test-fl_LoadFjord.R
@@ -1,4 +1,4 @@
-# Testing for fl_ListFjords
+# Testing for fl_LoadFjords
 
 test_that("loading errors are correct", {
   expect_error(fl_LoadFjord("test"),
@@ -8,12 +8,11 @@ test_that("loading errors are correct", {
 })
 
 test_that("loading with and without TS works", {
-  fl_DownloadFjord(fjord = "test", tempdir())
-  test_NO_TS <- fl_LoadFjord("test", dirdata = tempdir())
+  test_NO_TS <- fl_LoadFjord("test", system.file("extdata", package = "FjordLight"))
   expect_type(test_NO_TS, "list")
   expect_equal(length(test_NO_TS$longitude), 7)
   expect_equal(round(test_NO_TS$AreaOfShallowZone), 106)
-  test_TS <- fl_LoadFjord("test", dirdata = tempdir(), TS = TRUE)
+  test_TS <- fl_LoadFjord("test", dirdata = system.file("extdata", package = "FjordLight"), TS = TRUE)
   expect_type(test_TS, "list")
   expect_type(test_TS$MonthlyPARbottom, "double")
 })