+++ tests/testthat/test-flget_climatology.R
@@ -2,8 +2,7 @@
 # This also tests flplot_climatology
 
 test_that("flget_climatology functions correctly", {
-  fl_DownloadFjord(fjord = "test", tempdir())
-  dat_no_TS <- fl_LoadFjord("test", tempdir())
+  dat_no_TS <- fl_LoadFjord("test", system.file("extdata", package = "FjordLight"))
 
   expect_s4_class(flget_climatology(dat_no_TS), "RasterLayer")
 
@@ -22,9 +21,9 @@ test_that("flget_climatology functions correctly", {
   expect_error(flget_climatology(dat_no_TS, period = "Clim", month = 6:8),
                "Please select a single month")
   expect_error(flget_climatology(dat_no_TS, period = "Clim", month = 1),
-               "Bad month: available months 3 4 5 6 7 8 9 10")
+               "Bad month: available months 3 4 5 6 7 8 9")
   expect_error(flget_climatology(dat_no_TS, period = "Clim", month = "banana"),
-               "Bad month: available months 3 4 5 6 7 8 9 10")
+               "Bad month: available months 3 4 5 6 7 8 9")
 
   expect_error(flget_climatology(dat_no_TS, period = "Yearly"),
                "Please indicate the year")