+++ tests/testthat/test-flget_geoparameters.R
@@ -1,8 +1,7 @@
 # Tests for flget_geoparameters
 
 test_that("flget_geoparameters works", {
-  fl_DownloadFjord(fjord = "test", tempdir())
-  dat_no_TS <- fl_LoadFjord("test", dirdata = tempdir())
+  dat_no_TS <- fl_LoadFjord("test", dirdata = system.file("extdata", package = "FjordLight"))
   res_geo <- flget_geoparameters(dat_no_TS)
   expect_type(res_geo, "double")
 })