+++ NEWS.md
@@ -1,3 +1,16 @@
+# FjordLight 0.7.0 (2023-11-30)
+
+* Functions now detect the available years and months of data within the dataset and warn users accordingly
+
+# FjordLight 0.6.2 (2023-11-30)
+
+* Changed testing structure to function correctly when offline
+* Added `test.nc` as a bonus file in 'inst/extdata/'
+
+# FjordLight 0.6.2 (2023-11-15)
+
+* Changed testing structure to prevent future CRAN ERROR issues if FTP structure is changed
+
 # FjordLight 0.6.1 (2023-11-13)
 
 * Change to download function to allow downloading over slow connection speed